1318 - move layout folder to client
