+++ src/components/client/layout/Layout.tsx
@@ -4,10 +4,10 @@ import { useMemo, useState } from 'react'
 import { useTranslation } from 'next-i18next'
 import { Box, BoxProps, Container, ContainerProps, Typography } from '@mui/material'
 
-import Footer from 'components/layout/Footer/Footer'
+import Footer from 'components/client/layout/Footer/Footer'
+import Snackbar from 'components/client/layout/Snackbar'
 import { defaultOgImage } from 'common/routes'
-import Snackbar from 'components/layout/Snackbar'
-import DetailsModal from 'components/modal/DetailsModal'
+import DetailsModal from 'components/admin/modal/DetailsModal'
 
 import AppNavBar from './AppNavBar'
 import MobileNav from './nav/MobileNav/MobileNav'
