+++ src/components/client/layout/NotificationSnackBar/NotificationSnackBar.tsx
