+++ src/components/client/layout/NotificationSnackBar/props/global.ts
